+++ kwimage/im_draw.py
@@ -1091,6 +1091,25 @@ def fill_nans_with_checkers(canvas, square_shape=8):
         >>> kwplot.autompl()
         >>> kwplot.imshow(img, pnum=(1, 2, 1))
         >>> kwplot.imshow(canvas, pnum=(1, 2, 2))
+
+    Example:
+        >>> # Test grayscale
+        >>> from kwimage.im_draw import *  # NOQA
+        >>> import kwimage
+        >>> orig_img = kwimage.ensure_float01(kwimage.grab_test_image())
+        >>> poly1 = kwimage.Polygon.random().scale(orig_img.shape[0] // 2)
+        >>> poly2 = kwimage.Polygon.random().scale(orig_img.shape[0])
+        >>> img = orig_img.copy()
+        >>> img = poly1.fill(img, np.nan)
+        >>> img[:, :, 0] = poly2.fill(np.ascontiguousarray(img[:, :, 0]), np.nan)
+        >>> img = kwimage.convert_colorspace(img, 'rgb', 'gray')
+        >>> canvas = img.copy()
+        >>> canvas = fill_nans_with_checkers(canvas)
+        >>> # xdoc: +REQUIRES(--show)
+        >>> import kwplot
+        >>> kwplot.autompl()
+        >>> kwplot.imshow(img, pnum=(1, 2, 1))
+        >>> kwplot.imshow(canvas, pnum=(1, 2, 2))
     """
     invalid_mask = np.isnan(canvas)
     return _masked_checkerboard(canvas, invalid_mask, square_shape)
@@ -1100,6 +1119,7 @@ def _masked_checkerboard(canvas, invalid_mask, square_shape):
     import kwimage
     import kwarray
     canvas = kwarray.atleast_nd(canvas, 3)
+    invalid_mask = kwarray.atleast_nd(invalid_mask, 3)
     allchan_invalid_mask = invalid_mask.all(axis=2, keepdims=1)
     anychan_invalid_mask = invalid_mask.any(axis=2, keepdims=1)